Reorganize components.  Use functional components where class components are not needed.
